Moved KuraBirthPayload, KuraDeviceProfile and KuraDisconnectPayload to APIs.
