+++ kura/org.eclipse.kura.api/META-INF/MANIFEST.MF
@@ -42,7 +42,7 @@ Export-Package: org.eclipse.kura;version="1.5.0",
  org.eclipse.kura.gpio;version="1.1.0",
  org.eclipse.kura.linux.udev;version="1.0.1",
  org.eclipse.kura.marshalling;version="1.0.0",
- org.eclipse.kura.message;version="1.2.0",
+ org.eclipse.kura.message;version="1.3.0",
  org.eclipse.kura.net;version="2.0.0",
  org.eclipse.kura.net.dhcp;version="1.1.0",
  org.eclipse.kura.net.dns;version="1.1.0",
