+++ kura/org.eclipse.kura.cloudconnection.eclipseiot.mqtt.provider/src/main/java/org/eclipse/kura/internal/cloudconnection/eclipseiot/mqtt/cloud/LifeCyclePayloadBuilder.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2011, 2018 Eurotech and/or its affiliates
+ * Copyright (c) 2011, 2019 Eurotech and/or its affiliates
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -14,10 +14,10 @@ package org.eclipse.kura.internal.cloudconnection.eclipseiot.mqtt.cloud;
 import java.util.List;
 
 import org.eclipse.kura.core.util.NetUtil;
-import org.eclipse.kura.internal.cloudconnection.eclipseiot.mqtt.message.KuraBirthPayload;
-import org.eclipse.kura.internal.cloudconnection.eclipseiot.mqtt.message.KuraDeviceProfile;
-import org.eclipse.kura.internal.cloudconnection.eclipseiot.mqtt.message.KuraDisconnectPayload;
-import org.eclipse.kura.internal.cloudconnection.eclipseiot.mqtt.message.KuraBirthPayload.KuraBirthPayloadBuilder;
+import org.eclipse.kura.message.KuraBirthPayload;
+import org.eclipse.kura.message.KuraBirthPayload.KuraBirthPayloadBuilder;
+import org.eclipse.kura.message.KuraDeviceProfile;
+import org.eclipse.kura.message.KuraDisconnectPayload;
 import org.eclipse.kura.message.KuraPosition;
 import org.eclipse.kura.net.NetInterface;
 import org.eclipse.kura.net.NetInterfaceAddress;
@@ -60,7 +60,8 @@ public class LifeCyclePayloadBuilder {
             deviceName = this.cloudConnectionManagerImpl.getSystemService().getDeviceName();
         }
 
-        String payloadEncoding = this.cloudConnectionManagerImpl.getCloudConnectionManagerOptions().getPayloadEncoding().name();
+        String payloadEncoding = this.cloudConnectionManagerImpl.getCloudConnectionManagerOptions().getPayloadEncoding()
+                .name();
 
         // build birth certificate
         KuraBirthPayloadBuilder birthPayloadBuilder = new KuraBirthPayloadBuilder();
