+++ kura/org.eclipse.kura.core.cloud/src/main/java/org/eclipse/kura/core/cloud/LifeCyclePayloadBuilder.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2011, 2017 Eurotech and/or its affiliates
+ * Copyright (c) 2011, 2019 Eurotech and/or its affiliates
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -13,11 +13,11 @@ package org.eclipse.kura.core.cloud;
 
 import java.util.List;
 
-import org.eclipse.kura.core.message.KuraBirthPayload;
-import org.eclipse.kura.core.message.KuraBirthPayload.KuraBirthPayloadBuilder;
-import org.eclipse.kura.core.message.KuraDeviceProfile;
-import org.eclipse.kura.core.message.KuraDisconnectPayload;
 import org.eclipse.kura.core.util.NetUtil;
+import org.eclipse.kura.message.KuraBirthPayload;
+import org.eclipse.kura.message.KuraBirthPayload.KuraBirthPayloadBuilder;
+import org.eclipse.kura.message.KuraDeviceProfile;
+import org.eclipse.kura.message.KuraDisconnectPayload;
 import org.eclipse.kura.message.KuraPosition;
 import org.eclipse.kura.net.NetInterface;
 import org.eclipse.kura.net.NetInterfaceAddress;
