+++ kura/test/org.eclipse.kura.core.cloud.test/META-INF/MANIFEST.MF
@@ -8,6 +8,7 @@ Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ClassPath: .
 Bundle-ActivationPolicy: lazy
 Import-Package: org.eclipse.kura.core.testutil,
+ org.eclipse.kura.message;version="1.2.0",
  org.eclipse.kura.test.annotation;version="1.0.0",
  org.junit;version="4.12.0",
  org.junit.runners;version="4.12.0",
