+++ kura/test/org.eclipse.kura.core.cloud.test/src/test/java/org/eclipse/kura/core/message/KuraBirthPayloadTest.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2017 Eurotech and/or its affiliates and others
+ * Copyright (c) 2017, 2019 Eurotech and/or its affiliates and others
  *
  *   All rights reserved. This program and the accompanying materials
  *   are made available under the terms of the Eclipse Public License v1.0
@@ -12,7 +12,8 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-import org.eclipse.kura.core.message.KuraBirthPayload.KuraBirthPayloadBuilder;
+import org.eclipse.kura.message.KuraBirthPayload;
+import org.eclipse.kura.message.KuraBirthPayload.KuraBirthPayloadBuilder;
 import org.eclipse.kura.message.KuraPosition;
 import org.junit.Test;
